Moved TrainPipeline to train.py
